+++ examples/fallen showcase/main.js
@@ -1,35 +1,24 @@
 import { Elements } from "/ruitk/elements/elements.js";
 import { TileWin } from "/ruitk/tileWin/tileWin.js";
-import { Theme } from "/ruitk/themes/fallen/main.js";
-
-let app = document.createElement("img");
-app.src = "/ruitk/themes/fallen/wallpaper.jpg";
-app.style.position = "fixed";
-app.style.left = "0px";
-app.style.top = "0px";
-app.style.width = "100%";
-app.style.height = "100%";
-app.style.objectFit = "cover";
-app.style.zIndex = "-2";
-
-document.querySelector("body").appendChild(app);
-
+import { FallenBase } from "/ruitk/themes/fallen/base.js";
 
 let tileWin = new TileWin();
 let elements = new Elements();
-let theme = new Theme();
+FallenBase.init();
 
+tileWin.updateStyle(FallenBase.getTileWinStyle());
+elements.addElements(FallenBase.getElements());
 
-tileWin.updateConfig({
-    animateOnCreateTile : true,
-    createInnerTile : true,
-});
-
-tileWin.updateStyle(theme.getTileWinStyle());
-elements.addElements(theme.getElements());
+elements.append(elements.makeElements(`<backgroundImg>{}`), "body");
 
 tileWin.createTile("TextWindow", 1, 0, 1, 0, elements.makeElements(`
-    <p1>{"text" : "p1 p1 p1 p1 p1 p1 p1"}
+    <h1>{"content" : "h1 h1 h1 h1 h1 h1 h1"}
+    <h2>{"content" : "h2 h2 h2 h2 h2 h2 h2"}
+    <h3>{"content" : "h3 h3 h3 h3 h3 h3 h3"}
+
+    <p1>{"content" : "p1 p1 p1 p1 p1 p1 p1 " <b>{"content" : "Bold", "textAline : "right"}}
+    <p2>{"content" : "p2 p2 p2 p2 p2 p2 p2"}
+    <p3>{"content" : "p3 p3 p3 p3 p3 p3 p3"}
     `));
 tileWin.createTile("Tile the 2nd",2,0,1,0);
 tileWin.update();